% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ancestry_chunks}
\alias{ancestry_chunks}
\title{Chunks}
\usage{
ancestry_chunks(jumps)
}
\arguments{
\item{jumps}{a numeric vector obtained with \code{jumps_from_map}.}
}
\value{
The returned value is a logical vector.
}
\description{
\code{ancestry_chunks} converts jumps to chunks.
}
