% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{calculate() Function}
\usage{
calculate(equations = NULL, setx = NULL, newdata = NULL, keep = FALSE)
}
\arguments{
\item{equations}{(Required) Data.frame containing multivariate normal distribution of coefficient terms, generated by equate().}

\item{setx}{(Optional) List containing specific hypothetical predictor values of X for key coefficients. All others held at medians/modes.}

\item{newdata}{(Optional) Data.frame containing hypothetical predictor values of X for K-columns of coefficients.}

\item{keep}{(FALSE by default) Logical which if TRUE tells model to keep the contents of newdata in output data.frame. If FALSE, exclude them for efficiency.}
}
\description{
This function calculates simulated outcomes based on an N-row data.frame of K-columns of coefficients, supplied from equate().
}
\examples{

# Load Packages
library(simulate)
library(tidyverse)

# Make a regression model
mymodel <- mtcars \%>\% lm(formula = mpg ~ disp + factor(cyl))

# Generate 1000 simulations, accounting for estimation uncertainty
mymodel \%>\%
  equations() \%>\%
  calculate(newdata = mtcars)

# or...

mymodel \%>\%
  equate() \%>\%
  calculate(setx = list(disp = c(100, 200, 300)))
}
\keyword{simulation}
