% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{check_intersect}
\alias{check_intersect}
\title{Simulation tools}
\usage{
check_intersect(p, ground.truth)
}
\arguments{
\item{p}{a list containing two integers.}

\item{ground.truth}{a vector of integers.}
}
\value{

}
\description{
\code{check_intersect}
}
