% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{cor_matrix_str}
\alias{cor_matrix_str}
\title{Input structure for correlation matrix}
\usage{
cor_matrix_str(n)
}
\arguments{
\item{n}{The dimension of the correlation matrix.}
}
\description{
This function is designed for interactive use to quickly generate the
input structure needed for \code{correlation_matrix} function.
}
\seealso{
Other correlation matrix: 
\code{\link{correlation_matrix}()}
}
\concept{correlation matrix}
