% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Generate correlated random variables}
\usage{
correlate(
  .data,
  vars = NULL,
  mean = NULL,
  cor = NULL,
  cov = NULL,
  dist = "normal"
)
}
\description{
Generate correlated random variables
}
