% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{create_all_inputs}
\alias{create_all_inputs}
\title{Input tools}
\usage{
create_all_inputs(H1, H2, H, directory, pos_map, gen_map, pop, df, gt.df)
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}

\item{H}{a haplotype matrix.}

\item{directory}{a character string.}

\item{pos_map}{a numeric vector.}

\item{gen_map}{a numeric vector.}

\item{pop}{a numeric vector.}

\item{df}{a data.frame.}

\item{gt.df}{a data.frame.}
}
\value{

}
\description{
\code{create_all_inputs}
}
