% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-EILA.R
\name{create_input_EILA}
\alias{create_input_EILA}
\title{Input tools}
\usage{
create_input_EILA(H1, H2, Hadm, position)
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}

\item{Hadm}{a haplotype matrix.}

\item{position}{a numerical vector.}
}
\value{

}
\description{
\code{create_input_EILA} creates an object storing the input matrices 
required by \code{EILA::eila}.
}
