% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-Hapmix.R
\name{create_input_Hapmix}
\alias{create_input_Hapmix}
\title{Input tools}
\usage{
create_input_Hapmix(H1, H2, Gadm, gen_map, phys_map, chr, Hapmix.output,
  ancestral.1 = "ANC1", ancestral.2 = "ANC2", admixed = "AA",
  rates.digits = 8)
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}

\item{Gadm}{a genotype matrix.}

\item{gen_map}{a numerical vector.}

\item{phys_map}{a numerical vector.}

\item{chr}{an integer.}

\item{Hapmix.output}{a character string.}

\item{ancestral.1}{a character string.}

\item{ancestral.2}{a character string.}

\item{admixed}{a character string.}

\item{rates.digits}{an integer.}
}
\value{

}
\description{
\code{create_input_Hapmix} generates input files for Hapmix.
}
