% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-loter.R
\name{create_input_loter}
\alias{create_input_loter}
\title{Input tools}
\usage{
create_input_loter(H1, H2, Hadm, loter.output)
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}

\item{Hadm}{a haplotype matrix.}

\item{loter.output}{a character string.}
}
\value{

}
\description{
\code{create_input_loter}
}
