% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{create_input_pcadapt}
\alias{create_input_pcadapt}
\title{Input tools}
\usage{
create_input_pcadapt(H1, H2, H)
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}

\item{H}{a haplotype matrix.}
}
\value{

}
\description{
\code{create_input_pcadapt} creates a genotype matrix in the \code{pcadapt} 
format.
}
