% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-RASPberry.R
\name{create_input_RASPberry}
\alias{create_input_RASPberry}
\title{Input tools}
\usage{
create_input_RASPberry(refpop1_phased, refpop2_phased, H, pos_map, ref, alt,
  chr)
}
\arguments{
\item{refpop1_phased}{a haplotype matrix.}

\item{refpop2_phased}{a haplotype matrix.}

\item{H}{a genotype matrix.}

\item{pos_map}{a numerical vector.}

\item{ref}{a string vector.}

\item{alt}{a string vector.}

\item{chr}{an integer.}
}
\description{
\code{create_input_RASPberry} generates input files for RASPberry.
}
