% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-RFMix.R
\name{create_input_RFMix}
\alias{create_input_RFMix}
\title{RFMix}
\usage{
create_input_RFMix(H1, H2, Hadm, genetic.map, RFMix.output = "RFMix")
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}

\item{Hadm}{a haplotype matrix.}

\item{genetic.map}{a numerical vector.}

\item{RFMix.output}{a character string.}
}
\value{

}
\description{
\code{create_input_RFMix}
}
