% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detect_cpp}
\alias{detect_cpp}
\title{Peak detection}
\usage{
detect_cpp(stat, threshold)
}
\arguments{
\item{stat}{a numeric vector.}

\item{tol}{an integer.}
}
\value{
The returned value is a numeric vector.
}
\description{
\code{detect_cpp} returns a numeric vector determining where switches
are located.
}
