% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry.R
\name{display.ancestry}
\alias{display.ancestry}
\title{Ancestry tools}
\usage{
display.ancestry(ancestry.matrix)
}
\arguments{
\item{rfmix.output}{a character string.}
}
\value{

}
\description{
\code{display.ancestry}
}
