% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-EILA.R
\name{EILA_from_pcadapt}
\alias{EILA_from_pcadapt}
\title{Input tools}
\usage{
EILA_from_pcadapt(geno, pop, ancestral.1, ancestral.2, admixed, position)
}
\arguments{
\item{geno}{a genotype matrix.}

\item{pop}{a vector of integers.}

\item{ancestral.1}{an integer.}

\item{ancestral.2}{an integer.}

\item{admixed}{an integer.}

\item{position}{a numerical vector.}
}
\value{

}
\description{
\code{EILA_from_pcadapt} creates an object storing the input matrices 
required by \code{EILA::eila}.
}
