% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate.R
\name{equate}
\alias{equate}
\title{equate() Function}
\usage{
equate(m = NULL, reps = 1000, seed = 12345, package = "mvtnorm")
}
\arguments{
\item{m}{Model object. Required.}

\item{reps}{Number of simulations to generate of coefficients. Defaults to 1000.}

\item{seed}{Seed for perfect replication of simulations. Defaults to 12345 unless otherwise specified.}

\item{package}{Package used to simulate. Options include \code{mvtnorm} (recommended) or \code{MASS}.}
}
\description{
This function allows you to generate a multivariate normal distribution
}
\examples{

# Load Packages
library(simulate)
library(tidyverse)

# Make a regression model
mymodel <- mtcars \%>\% lm(formula = mpg ~ disp + factor(cyl))

# Generate a multivariate normal distribution for a set of model coefficients.
d <- equate(m = mymodel)

# Check out the first 5 lines below. Notice the unique ID for each row, listed under replicate.
head(d)

}
\keyword{simulate}
