% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_treatment_assignments.R
\name{gen_trt}
\alias{gen_trt}
\title{Generating random treatment assignments}
\usage{
gen_trt(formula, data, beta, indep = FALSE, trt.frac = 0.5, n)
}
\arguments{
\item{formula}{an object of class \code{"formula"}, a symbolic description of the model
for the logistic model for treatment assignments (the propensity model)}

\item{data}{data.frame of covariates for propensity model}

\item{beta}{vector of covariate effects for the logistic propensity model. Must be equal
in length to the design matrix generated by the specified \code{formula}}

\item{indep}{logical value. If \code{FALSE} the treatment assignments will be generated by
the model specified by \code{formula}; if \code{TRUE}, treatment assignments will be generated
indepenedently of any covariates, with the treatment taking value \code{1} on average
\code{trt.frac} * 100 percent of the time. User must specify the sample size via
\code{n} if \code{indep = TRUE}.}

\item{trt.frac}{fraction (between 0 and 1) of treatment values that take value \code{1}}

\item{n}{positive integer value for sample size. Must be specified if \code{indep = TRUE}.}
}
\description{
Generates random treatment assignments from a specified logistic model
}
\examples{

beta <- gen_coefs(20, nonzero.frac = 0.5,
                  max.effect.size = 0.5)
beta
}
