% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_hybrid}
\alias{generate_hybrid}
\title{Simulation tools}
\usage{
generate_hybrid(H1, H2, map, n.gen = 10, global.ancestry = 0.5)
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}

\item{map}{a numerical vector containing the genetic positions.}

\item{global.ancestry}{a numerical value between \code{0} and \code{1}
setting the average proportion of markers coming from the \code{H1}.}
}
\value{

}
\description{
\code{generate_hybrid} generates one hybrid individual using the 
recombination map and ancestral haplotypes.
}
