% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_hybrid_matrix}
\alias{generate_hybrid_matrix}
\title{Simulation tools}
\usage{
generate_hybrid_matrix(H1, H2, alpha = 0.5, beta = 1 - alpha, gen_map,
  n.hyb, lambda = 1, ancestry.switch = NULL)
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}

\item{alpha}{a numerical value.}

\item{beta}{a numerical value.}

\item{gen_map}{a numerical vector.}

\item{n.hyb}{an integer.}

\item{lambda}{a numerical value.}

\item{ancestry.switch}{a numerical vector.}

\item{jumps}{a numerical vector.}
}
\value{

}
\description{
\code{generate_hybrid_matrix} generates hybrid individuals using the 
recombination map and ancestral haplotypes.
}
