% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coeftable.R
\name{get_coeftable}
\alias{get_coeftable}
\title{get_coeftable() Function}
\format{
\describe{
\item{term}{name of variable for which estimation uncertainty was measured}
\item{estimate}{observed coefficient}
\item{lower}{lower confidence interval for simulated coefficients, based on input ci. When ci = 0.95, defaults to 2.5th percentile.}
\item{upper}{upper confidence interval for simulated coefficients, based on input ci. When ci = 0.95, defaults to 97.5th percentile.}
\item{se}{standard error, representing standard deviation of simulated coefficients.}
\item{p_value}{p-value calculated from simulations. Depicts false positive rate as the percentage of simulations on opposite side of 0 (mu) from the observed coefficient.}
\item{stars}{Stars show statistical significance (*** p < 0.001, ** p < 0.01, * p < 0.05, . p < 0.10.)}
}
}
\usage{
get_coeftable(data, ci = 0.95, mu = 0)
}
\arguments{
\item{data}{Data.frame containing multivariate normal distribution of coefficient terms, generated by equate().}

\item{ci}{(Optional) Single value from 0 to 1. Defaults to 0.95, for a 95\% confidence interval.}

\item{mu}{(default = 0) To compute statistical significance of estimates, specify the null hypothesis (eg. estimate = 0). mu = 0 is most common.}
}
\description{
This underlying function derives a summary table based on basic alpha & coefficient standard errors, test statistics, and p-values from a multivariate normal distribution, circumventing constraints on ordinary standard errors.
}
\examples{

# Load Packages
library(simulate)
library(tidyverse)

# Make a regression model
mymodel <- mtcars \%>\% lm(formula = mpg ~ disp + factor(cyl))

# Generate simulated coefficient table
mymodel \%>\%
  equate() \%>\%
  get_coeftable()
  
# More recommended is to use the tabulate() function, 
# which provides the right output for each layer in the simulation process.
mymodel \%>\%
  equate() \%>\%
  tabulate()
  
}
\keyword{coefficients}
\keyword{summary}
\keyword{table}
