% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_dist_pop}
\alias{get_dist_pop}
\title{Simulation tools}
\usage{
get_dist_pop(H1, H2)
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}
}
\value{

}
\description{
\code{get_dist_pop}
}
