% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{get_mode() Function}
\usage{
get_mode(x)
}
\arguments{
\item{x}{A numeric, categorical, or factor vector.}
}
\description{
This function grabs the modal response from a vector.
}
\examples{

c("cat", "dog", "dog", "mouse", NA_character_) \%>\%
   get_mode()

}
\keyword{mode}
