% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moments.R
\name{get_moments}
\alias{get_moments}
\title{get_moments() Function}
\usage{
get_moments(d = NULL, mu = NULL, var = NULL)
}
\arguments{
\item{d}{(Optional) A numeric vector. If supplied, used to calculate \code{mu} and \code{var}.}

\item{mu}{(Optional) A single numeric value, representing the mean of beta distribution. Requires \code{var} simultaneously.}

\item{var}{(Optional) A single numeric value, representing the variance of beta distribution. Requires \code{mu} simultaneously.}
}
\description{
This function uses the method of moments to approximate alpha and beta parameters in a beta-distributed simulation.
}
\keyword{moments}
