% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_newdata.R
\name{get_newdata}
\alias{get_newdata}
\title{get_newdata() Function}
\usage{
get_newdata(m, setx = NULL)
}
\arguments{
\item{m}{Model object. Required.}

\item{setx}{named list of hypothetical predictor values, for any variables of interest. All others set to median (for numeric variables) or mode (for categorical variables) by default. Note: This package currently cannot handle factor() language in model formulas; please transform any relevant variables to factors before modeling.}
}
\description{
This wrapper function extracts the median/modal observations for a set of data,
}
\examples{

# Load Packages
library(simulate)
library(tidyverse)

# Make a regression model
mymodel <- mtcars \%>\% lm(formula = mpg ~ disp + factor(cyl))

# Get a data.frame of the average respondent, if just the following variable values were changed
newdata <- get_newdata(m = mymodel, setx = list(disp = c(1,2,3)))

}
\keyword{average}
\keyword{mode}
