% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantiles.R
\name{get_quantiles}
\alias{get_quantiles}
\title{get_quantiles() Function}
\format{
\describe{
\item{term}{name of variable for which quantities of interest were calculated, eg. ev or pv.}
\item{estimate}{median quantity of interest}
\item{lower}{lower confidence interval for quantities of interest, based on input ci. When ci = 0.95, defaults to 2.5th percentile.}
\item{upper}{upper confidence interval for quantities of interest, based on input ci. When ci = 0.95, defaults to 97.5th percentile.}
\item{se}{standard error, representing standard deviation of quantities of interest.}
\item{p_value}{p-value calculated from simulations. Depicts false positive rate as the percentage of simulations on opposite side of mu (usually 0) from the median quantity of interest. p-values shown only if mu is specified. p-values are only interpretable when mu is meaningful, such as when calculating first differences where mu = 0.}
\item{stars}{Stars show statistical significance (*** p < 0.001, ** p < 0.01, * p < 0.05, . p < 0.10.)}
}
}
\usage{
get_quantiles(data, qi, ci = 0.95, mu = NULL)
}
\arguments{
\item{data}{(Required) Data.frame containing N predicted and expected values}

\item{qi}{(Required) Name of vector (length 1) in data frame to summary, written as a character vector. Typically ev, pv, or fd, but users can supply their own if needed.}

\item{ci}{(Optional) Single value from 0 to 1. Defaults to 0.95, for a 95\% confidence interval.}

\item{mu}{(Optional) To compute statistical significance, specify the null hypothesis (eg. first difference = 0). mu = 0 is most common.}
}
\description{
This underlying function derives a summary table based on predicted or expected values simulated by the simulate() function..
}
\examples{

# Load Packages
library(simulate)


}
\seealso{
\code{\link[=summary]{summary()}}, the wrapper function for summarizing all quantities of interest from this package, and \code{\link[=get_coeftable]{get_coeftable()}}, for regression coefficient estimates.
}
\keyword{expected}
\keyword{predicted}
\keyword{simulation}
\keyword{summary}
\keyword{table}
