% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim.R
\name{get_sim}
\alias{get_sim}
\title{get_sim() Function}
\usage{
get_sim(i, equations, newdata = newdata, keep = keep)
}
\arguments{
\item{i}{(Require) single number representing the row-number of the multivariate normal distribution simulated in equations().}

\item{equations}{(Required) Data.frame containing multivariate normal distribution of coefficient terms, generated by equations().}

\item{newdata}{(Required) Data.frame containing hypothetical predictor values of X for K-columns of coefficients.}

\item{keep}{(FALSE by default) Logical which if TRUE tells model to keep the contents of newdata in output data.frame. If FALSE, exclude them for efficiency.}
}
\description{
This function calculates simulated outcomes based on an N-row data.frame of K-columns of coefficients, supplied from equations(). Internal function; use get_predictions() instead.
}
\keyword{simulation}
