% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry.R
\name{haplo_to_ancestry}
\alias{haplo_to_ancestry}
\title{Ancestry tools}
\usage{
haplo_to_ancestry(H, anc)
}
\arguments{
\item{anc}{an integer.}

\item{H1}{a haplotype matrix.}
}
\value{

}
\description{
\code{haplo_to_ancestry} computes a vector of local ancestries from a 
haplotype matrix.
}
