% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{haplo_to_geno}
\alias{haplo_to_geno}
\title{Convert haplotypes to genotypes}
\usage{
haplo_to_geno(H)
}
\arguments{
\item{H}{a numerical matrix.}
}
\value{
The returned value is a numerical matrix.
}
\description{
\code{haplo_to_geno} converts haplotypes to genotypes.
}
