% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jumps_from_map}
\alias{jumps_from_map}
\title{Jumps}
\usage{
jumps_from_map(map, lambda)
}
\arguments{
\item{map}{a numeric vector.}

\item{lambda}{a numerical value specifying the parameter of the exponential
law.}
}
\value{
The returned value is a numeric vector.
}
\description{
\code{jumps_from_map} returns a numeric vector determining where jumps
are located.
}
