% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix-constructors}
\alias{matrix-constructors}
\alias{matrix_id}
\alias{matrix_diag}
\alias{matrix_cs}
\alias{matrix_ar1}
\alias{matrix_ad1}
\title{Constructors for structured correlation and covariance matrix}
\usage{
matrix_id(n, vars = 1, names = NULL, inverse_return = FALSE)

matrix_diag(n, vars = 1, names = NULL, inverse_return = FALSE)

matrix_cs(n, rho = 0, vars = 1, names = NULL, inverse_return = FALSE)

matrix_ar1(n, rho = 0, vars = 1, names = NULL, inverse_return = FALSE)

matrix_ad1(n, u, vars = 1, names = NULL, inverse_return = FALSE)
}
\arguments{
\item{n}{The dimension of the matrix.}

\item{vars}{A vector of diagonal variances or a single variance.}

\item{names}{Optional character vector of the same size as \code{n}.}

\item{inverse_return}{Whether to return the inverse or not.}

\item{rho}{The}
}
\description{
Correlation and covariance matrices are assumed to be symmetrics, semi-positive
definite. In the case of correlation matrices, the diagonal elements should be 1
and the off-diagonal elements between -1 and 1 inclusive.
}
\section{Functions}{
\itemize{
\item \code{matrix_id}: Identity matrix.

\item \code{matrix_diag}: A synonym for \code{matrix_id}.

\item \code{matrix_cs}: Compound symmetric matrix.

\item \code{matrix_ar1}: First order autoregressive order matrix.

\item \code{matrix_ad1}: First order ante-dependence order matrix.
}}

