% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrixop}
\alias{matrixop}
\alias{mset_values}
\alias{mset_matrix}
\title{Set the values of the cell in subset of matrix.}
\usage{
mset_values(matrix, i, j, replacement)

mset_matrix(matrix, i, j, replacement)
}
\arguments{
\item{matrix}{A matrix.}

\item{i, j}{The index of the rows and columns respectively.}

\item{replacement}{The values to set the}
}
\description{
If replacing specific cell values use \code{mset_values}.
If you want to change a sub-matrix of the whole matrix then use \code{mset_matrix}.
}
\section{Functions}{
\itemize{
\item \code{mset_values}: Set the values of particular cells

\item \code{mset_matrix}: Set the values of a submatrix
}}

\examples{

}
\seealso{
Other Matrix manipulators: 
\code{\link{mset_names}()}
}
\concept{Matrix manipulators}
