% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{mset_names}
\alias{mset_names}
\alias{mset_colnames}
\alias{mset_rownames}
\title{Set the name of the matrix.}
\usage{
mset_names(X, names = NULL)

mset_colnames(X, names = NULL)

mset_rownames(X, names = NULL)
}
\arguments{
\item{X}{A matrix.}

\item{names}{A character vector corresponding to the variable names. For
\code{mset_names}, it can also be a list of two character vectors with the
first and second elements corresponding to row and column names
respectively.}
}
\description{
Set the name of the matrix.
}
\section{Functions}{
\itemize{
\item \code{mset_colnames}: Set column names only.

\item \code{mset_rownames}: Set row names only.
}}

\seealso{
Other Matrix manipulators: 
\code{\link{mset_values}()}
}
\concept{Matrix manipulators}
