% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{mset_values}
\alias{mset_values}
\title{Set the name of the matrix.}
\usage{
mset_values(X, i, j, value)
}
\arguments{
\item{X}{A matrix.}
}
\description{
Set the name of the matrix.
}
\seealso{
Other Matrix manipulators: 
\code{\link{mset_names}()}
}
\concept{Matrix manipulators}
