% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{number.to.string}
\alias{number.to.string}
\title{Utils}
\usage{
number.to.string(n)
}
\arguments{
\item{n}{a numeric value.}
}
\value{

}
\description{
\code{number.to.string} converts a numeric value to a character string.
}
