% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{params}
\alias{params}
\title{Set the parameters in a simulation framework}
\usage{
params(.x, .param = NULL, ..., .reset = FALSE)
}
\arguments{
\item{...}{A named list of values to set.}

\item{.reset}{To reset the values of the parameters.}

\item{x}{A simulation framework}

\item{param}{A character vector with the name of the parameter to modify or set.}
}
\description{
Set the parameters in a simulation framework
}
