% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_xxx.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{.simulateFun}
\title{sim_xxx}
\arguments{
\item{nsim}{number of response vectors to simulate; defaults to 1}

\item{pred}{predictions, generated with \code{predict(object, type = "link")}}

\item{linkinv}{inverse link function.}

\item{distr}{error distribution for the specific family}

\item{rand.eff}{vector of random effects; defaults to NA = no random 
effects}

\item{corStruct}{matrix for the temporal or spatial correlation structure;
defaults to NA = no correlation. See \code{\link{nlme::corClasses}}.}
}
\description{
Internal function to compute new predictions
}
\details{
Simulate new response data (so far for gam::mgcv)
}
\author{
Loic Chalmandrier, Florian Hartig, Severin Hauenstein, Lionel Hertzog, Mira Kattwinkel
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lme4}{\code{\link[lme4]{.simulateFun}}}
}}

