% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry.R
\name{rfmix.local.ancestry}
\alias{rfmix.local.ancestry}
\title{Ancestry tools}
\usage{
rfmix.local.ancestry(rfmix.output)
}
\arguments{
\item{rfmix.output}{a character string.}
}
\value{

}
\description{
\code{rfmix.local.ancestry} computes a vector of local ancestries from an
output file obtained with RFMix.
}
