% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro_reg.R
\name{sample_intro_reg}
\alias{sample_intro_reg}
\title{Sample introgression regions}
\usage{
sample_intro_reg(nSNP, n.reg = 5, intro.size)
}
\arguments{
\item{nSNP}{an integer.}

\item{n.reg}{an integer.}

\item{intro.size}{an integer.}
}
\value{

}
\description{
\code{sample_intro_reg} returns a list of introgressed alleles.
}
