% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro_reg.R
\name{sample_introgression_regions}
\alias{sample_introgression_regions}
\title{Sample introgression regions}
\usage{
sample_introgression_regions(nSNP, n.windows, n.regions)
}
\arguments{
\item{nSNP}{an integer.}

\item{n.windows}{an integer.}

\item{n.regions}{an integer.}
}
\value{

}
\description{
\code{sample_introgression_regions} returns a list of introgressed alleles.
}
