% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{sim_gamma}
\alias{sim_gamma}
\title{Simulation framework for Gamma random variable}
\usage{
sim_gamma(shape, rate, params = NULL, data = NULL)
}
\arguments{
\item{shape}{The shape.}

\item{rate}{The rate.}

\item{params}{The list of parameters.}

\item{data}{An optional data frame.}
}
\description{
Simulation framework for Gamma random variable
}
\seealso{
Other simulation frameworks: 
\code{\link{sim_bernoulli}()},
\code{\link{sim_beta}()},
\code{\link{sim_binomial}()},
\code{\link{sim_cauchy}()},
\code{\link{sim_chisq}()},
\code{\link{sim_exponential}()},
\code{\link{sim_form}()},
\code{\link{sim_f}()},
\code{\link{sim_geometric}()},
\code{\link{sim_hypergeometric}()},
\code{\link{sim_multinominal}()},
\code{\link{sim_negative_binomial}()},
\code{\link{sim_normal}()},
\code{\link{sim_poisson}()},
\code{\link{sim_t}()},
\code{\link{sim_uniform}()},
\code{\link{sim_weibull}()}
}
\concept{simulation frameworks}
