% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{sim_normal}
\alias{sim_normal}
\title{Simulation framework for Normal random variable}
\usage{
sim_normal(mean = 0, sd = 1, params = NULL, data = NULL)
}
\arguments{
\item{mean}{The mean.}

\item{sd}{The standard deviation.}

\item{params}{The list of parameters.}

\item{data}{An optional data frame.}
}
\description{
Simulation framework for Normal random variable
}
\seealso{
Other simulation frameworks: 
\code{\link{sim_bernoulli}()},
\code{\link{sim_beta}()},
\code{\link{sim_binomial}()},
\code{\link{sim_cauchy}()},
\code{\link{sim_chisq}()},
\code{\link{sim_exponential}()},
\code{\link{sim_form}()},
\code{\link{sim_f}()},
\code{\link{sim_gamma}()},
\code{\link{sim_geometric}()},
\code{\link{sim_hypergeometric}()},
\code{\link{sim_multinominal}()},
\code{\link{sim_negative_binomial}()},
\code{\link{sim_poisson}()},
\code{\link{sim_t}()},
\code{\link{sim_uniform}()},
\code{\link{sim_weibull}()}
}
\concept{simulation frameworks}
