% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{sim_poisson}
\alias{sim_poisson}
\title{Simulation framework for Poisson random variable}
\usage{
sim_poisson(lambda, params = NULL, data = NULL)
}
\arguments{
\item{lambda}{The mean.}

\item{params}{The list of parameters.}

\item{data}{An optional data frame.}
}
\description{
Simulation framework for Poisson random variable
}
\seealso{
Other simulation frameworks: 
\code{\link{sim_bernoulli}()},
\code{\link{sim_beta}()},
\code{\link{sim_binomial}()},
\code{\link{sim_cauchy}()},
\code{\link{sim_chisq}()},
\code{\link{sim_exponential}()},
\code{\link{sim_form}()},
\code{\link{sim_f}()},
\code{\link{sim_gamma}()},
\code{\link{sim_geometric}()},
\code{\link{sim_hypergeometric}()},
\code{\link{sim_multinominal}()},
\code{\link{sim_negative_binomial}()},
\code{\link{sim_normal}()},
\code{\link{sim_t}()},
\code{\link{sim_uniform}()},
\code{\link{sim_weibull}()}
}
\concept{simulation frameworks}
