% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_xxx.R
\name{sim_xxx}
\alias{sim_xxx}
\title{sim_xxx}
\usage{
sim_xxx(nsim = 1, pred, linkinv, distr, rand.eff = NA, cor.str = NA)
}
\arguments{
\item{nsim}{number of response vectors to simulate; defaults to 1}

\item{pred}{predictions, generated with \code{predict(object, type = "link")}}

\item{linkinv}{inverse link function.}

\item{distr}{error distribution for the specific family}

\item{rand.eff}{vector of random effects; defaults to NA = no random 
effects}

\item{cor.str}{matrix for the temporal or spatial correlation structure;
defaults to NA = no correlation}
}
\description{
Internal function to compute new prediction
}
\details{
Simulate new response data (so far for gam::mgcv)
}
\author{
Lionel Hertzog
}
