% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simu.write.ancstrl}
\alias{simu.write.ancstrl}
\title{Simulation tools}
\usage{
simu.write.ancstrl(obj.vcfR, pop, ancestral.1, ancestral.2, recombinationRate,
  keep, output.name = "output")
}
\arguments{
\item{obj.vcfR}{an object of class vcfR obtained with a Beagle output.}

\item{pop}{a string vector specifying the population labels for 
each individual present in the VCF file.}

\item{ancestral.1}{a character string or an integer specifying the first 
ancestral population.}

\item{ancestral.2}{a character string or an integer specifying the second 
ancestral population.}

\item{recombinationRate}{a numerical value expressed in centiMorgans per 
Megabase specifying the mean recombination rate of the species.}

\item{keep}{a list specifying the markers to keep. If missing,
all markers will be kept.}

\item{output.name}{a character string specifying the name of the output map
file.}
}
\value{

}
\description{
\code{simu.write.ancstrl} creates haplotype matrices and genotype matrices 
from Beagle outputs.
}
