% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-package.R
\docType{package}
\name{simulate-package}
\alias{simulate}
\alias{simulate-package}
\title{simulate: Simulate variables}
\description{
Easy simulation
}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
\keyword{internal}
