% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.betareg.R
\name{simulate.betareg}
\alias{simulate.betareg}
\title{simulate.betareg}
\usage{
\method{simulate}{betareg}(object, nsim)
}
\description{
Get new response values for beta regression
}
\details{
The function simulate new response data
}
\author{
Lionel Hertzog
}
