% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.gam.R
\name{simulate.gam}
\alias{simulate.gam}
\title{simulate.gam.beta}
\usage{
\method{simulate}{gam}(object, nsim)
}
\description{
Simulate new response from a GAM (mgcv)
}
\details{
The function simulate new response data
}
\author{
Lionel Hertzog
}
