% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.nlme.R
\name{simulate.lme}
\alias{simulate.lme}
\title{simulate.lme}
\usage{
\method{simulate}{lme}(object)
}
\description{
Simulate new response from a lme object (nlme)
}
\details{
The function simulate new response data
}
\author{
Lionel Hertzog
}
