% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stat.by.window}
\alias{stat.by.window}
\title{Utils}
\usage{
stat.by.window(stat, n.windows, window.size)
}
\arguments{
\item{stat}{a numeric vector.}

\item{n.windows}{an integer specifying the number of windows.}
}
\value{

}
\description{
\code{stat.by.window} split a vector of statistics into a certain number
windows.
}
