% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substantiate.R
\name{substantiate}
\alias{substantiate}
\title{substantiate() Function}
\usage{
substantiate(
  list_equations,
  stat = "rmse",
  per = "model",
  terms = NULL,
  drop_terms = NULL,
  drop_pattern = NULL,
  ci = 0.95,
  mu = 0
)
}
\arguments{
\item{list_equations}{a list of 2 'equation' objects outputted from \code{\link[=equate]{equate()}}. Should be from two different models.}

\item{stat}{statistic to be calculated (either \code{"rmse"} or \code{"mpse"}). \code{"rmse"} returns root mean squared error (sigma) showing average difference between coefficients. \code{"mpse"} returns the square of \code{"rmse"}.}

\item{per}{level of tabulation (either \code{"model"} or \code{"terms"}). \code{"model"} returns statistic for change between all coefficients; \code{"terms"} returns statistic for change between each individual coefficient.}

\item{terms}{(Optional) character vector of coefficient names to include. (Defaults to all.)}

\item{drop_terms}{(Optional) character vector of coefficient names to exclude. (Defaults to none.)}

\item{drop_pattern}{(Optional) single character regex pattern of coefficient names to exclude. (Defaults to none.)}

\item{ci}{(numeric value) level of confidence interval (eg. 0.90, 0.95, 0.99, 0.999) (Defaults to \code{0.95}.) Used to \code{\link[=tabulate]{tabulate()}} summary statistics.}

\item{mu}{(numeric value) alternative hypothesis for \code{stat}. (Defaults to \code{0}.) Used to \code{\link[=tabulate]{tabulate()}} summary statistics.}
}
\description{
This function compares simulated estimates from 2 models, to determine whether shared coefficients changed significantly between model specifications.
}
\examples{

# Load Packages
library(simulate)
library(tidyverse)


# Let's use dplyr's delightful starwars dataset
data(starwars)

# Let's predict the height of characters based on their species (eg. Ewok?)
m1 <- starwars \%>\%
 lm(formula = height ~ species)
# Let's predict the height of characters based on their species and gender
m2 <- starwars \%>\%
 lm(formula = height ~ species + gender)

# Let's simulate 1000 model equations for each,
# and bind them into a list together.
mymodels <- list(
 m1 \%>\% equate(),
 m2 \%>\% equate())

# Get RMSE (average difference in coefficients' standardized test statistics)
# at the overall model level
mymodels \%>\%
 substantiate(per = "model")

# Get RMSE for each coefficeint
mymodels \%>\%
 substantiate(per = "terms")

# Get RMSE for each coefficient,
# but compare only Gungans and Ewoks
mymodels \%>\%
 substantiate(per = "terms", term = c("speciesGungan", "speciesEwok"))

# Get difference for each coefficient,
# but drop the intercept and Gungans
mymodels \%>\%
 substantiate(per = "terms", drop_terms = c("(Intercept)", "speciesGungan"))

# Get the difference for shared coefficients,
# but exclude coefficients that contain speciesGungan or speciesEwok
mymodels \%>\%
 substantiate(per = "model", drop_pattern = "Gungan|Ewok")
}
\keyword{simulate}
