% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{create_all_inputs}
\alias{create_all_inputs}
\title{Input tools}
\usage{
create_all_inputs(refpop1_integer, refpop2_integer, refpop1, refpop2, H.integer,
  H, true.ancestry.matrix, directory, pos_map, gen_map, chr, ref, alt, pop,
  parameters, gt.df)
}
\arguments{
\item{refpop1_integer}{a haplotype matrix.}

\item{refpop2_integer}{a haplotype matrix.}

\item{refpop1}{a haplotype matrix.}

\item{refpop2}{a haplotype matrix.}

\item{H.integer}{a haplotype matrix.}

\item{H}{a haplotype matrix.}

\item{true.ancestry.matrix}{a matrix.}

\item{directory}{a character string.}

\item{pos_map}{a numeric vector.}

\item{gen_map}{a numeric vector.}

\item{chr}{an integer.}

\item{ref}{a numeric vector.}

\item{alt}{a numeric vector.}

\item{pop}{a numeric vector.}

\item{parameters}{a data.frame.}

\item{gt.df}{a data.frame.}
}
\value{

}
\description{
\code{create_all_inputs}
}
