% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{create_input_eila}
\alias{create_input_eila}
\title{Input tools}
\usage{
create_input_eila(H1, H2, H, position)
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}

\item{H}{a haplotype matrix.}

\item{position}{a numerical vector.}
}
\value{

}
\description{
\code{create_input_eila} creates an object storing the input matrices 
required by \code{EILA::eila}.
}
