% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{create_input_hapmix}
\alias{create_input_hapmix}
\title{Input tools}
\usage{
create_input_hapmix(H1, H2, H, gen_map, phys_map, chr, ancstrl.1 = "ANC1",
  ancstrl.2 = "ANC2", admxd = "AA", rates.digits = 8)
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}

\item{H}{a haplotype matrix.}

\item{gen_map}{a numerical vector.}

\item{phys_map}{a numerical vector.}

\item{chr}{an integer.}

\item{ancstrl.1}{a character string.}

\item{ancstrl.2}{a character string.}

\item{admxd}{a character string.}

\item{rates.digits}{an integer.}
}
\value{

}
\description{
\code{create_input_hapmix} generates input files for Hapmix.
}
