% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{create_input_rfmix}
\alias{create_input_rfmix}
\title{RFMix}
\usage{
create_input_rfmix(H1, H2, H3, gen_map, rfmix.output = "rfmix")
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}

\item{H3}{a haplotype matrix.}

\item{gen_map}{a numerical vector.}

\item{rfmix.output}{a character string.}
}
\value{

}
\description{
\code{create_input_rfmix}
}
