% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{eila_from_pcadapt}
\alias{eila_from_pcadapt}
\title{Input tools}
\usage{
eila_from_pcadapt(geno, pop, anc1, anc2, admixed, position)
}
\arguments{
\item{geno}{a genotype matrix.}

\item{pop}{a vector of integers.}

\item{anc1}{an integer.}

\item{anc2}{an integer.}

\item{admixed}{an integer.}

\item{position}{a numerical vector.}
}
\value{

}
\description{
\code{eila_from_pcadapt} creates an object storing the input matrices 
required by \code{EILA::eila}.
}
