% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_one_hybrid}
\alias{generate_one_hybrid}
\title{Simulation tools}
\usage{
generate_one_hybrid(H1, H2, alpha, beta = 1 - alpha, jumps, ancestry.switch)
}
\arguments{
\item{H1}{a haplotype matrix.}

\item{H2}{a haplotype matrix.}

\item{alpha}{a numerical value.}

\item{beta}{a numerical value.}

\item{jumps}{a numerical vector.}

\item{ancestry.switch}{a numerical vector.}
}
\value{

}
\description{
\code{generate_hybrid} generates one hybrid individual using the 
recombination map and ancestral haplotypes.
}
