% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simu.write.ancstrl}
\alias{simu.write.ancstrl}
\title{Simulation tools}
\usage{
simu.write.ancstrl(vcf.file, pop.file, ancstrl.1, ancstrl.2, recombinationRate,
  keep, output.name = "output")
}
\arguments{
\item{vcf.file}{a character string specifying the name of a VCF file 
generated with Beagle containing phased haplotypes.}

\item{pop.file}{a character string specifying the name of a text file 
containing the population labels for each individual present in the VCF file.}

\item{ancstrl.1}{a character string or an integer specifying the first 
ancestral population.}

\item{ancstrl.2}{a character string or an integer specifying the second 
ancestral population.}

\item{recombinationRate}{a numerical value expressed in cM, specifying the
mean recombination rate of the species.}

\item{keep}{an integer specifying the number of markers to keep. If missing,
all markers will be kept.}

\item{output.name}{a character string specifying the name of the output map
file.}
}
\value{

}
\description{
\code{simu.write.ancstrl} creates haplotype matrices and genotype matrices 
from Beagle outputs.
}
