% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-package.R
\docType{package}
\name{simulate-package}
\alias{simulate}
\alias{simulate-package}
\title{simulate: Parametric simulation framework to generate complex multivariate and multilevel data}
\description{
Simulation functions describe and generate complex multivariate and multilevel data from a combination of common parameteric distributions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emitanaka/simulate}
  \item Report bugs at \url{https://github.com/emitanaka/simulate/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
\keyword{internal}
