% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.betareg.R
\name{simulate.betareg}
\alias{simulate.betareg}
\title{simulate.betareg}
\usage{
\method{simulate}{betareg}(object, nsim)
}
\description{
Get new response values for beta regression
}
\details{
The function simulate new response data
}
\examples{
library(simulate)
library(betareg)

#fit a model from ?betareg examples
data("GasolineYield", package = "betareg")
gy <- betareg(yield ~ batch + temp, data = GasolineYield)
summary(gy)

#get 100 new simulations
sims <- simulate(gy,nsim=100)

}
\author{
Lionel Hertzog
}
